% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg2arc.R
\name{deg2arc}
\alias{deg2arc}
\title{Equivalence between degrees and arcmin/arcsec}
\usage{
deg2arc(toConvert, units = "s", inverse = FALSE)
}
\arguments{
\item{toConvert}{A raster object or a numeric value or vector.}

\item{units}{Measurement units of the output when transformation is from degrees to arc (\code{inverse = FALSE}). When transformation is from arc to degrees (\code{inverse = TRUE}), this is the input units. It may be \code{"m"} for arcmin, or \code{"s"} for arcsec. By default it is arcsec.}

\item{inverse}{Indicates the direction of the transformation, from degrees to arc when \code{FALSE}, or from arc to degrees when \code{TRUE}, in this case \code{toConvert} must be numeric, no raster input is supported.}
}
\value{
A value or vector of values with the equivalences.
}
\description{
\code{deg2arc} converts degrees to their equivalents in arcmin or arcserc, and vice versa.
}
